Move pipeline test data to a generic location for reuse
